PR# 104 Improvements to Ansible based CI framework ------------------------------------------------------------------ - Moved the files which were specific to inventory & iometer roles   into a 'files' folder of the respective roles
- Updated the inventory role defaults/main.yml, tasks/main.yml &
  iometer role tasks.main.yml to consume files from {{role_path}}

- Modified the generate-inventory python script. Changes include :

  - Handling exceptions at relevant places

  - Validation of the machines.in input file for inventory generation

     - Check for field length in each line
     - Validate the IPV4 addresses provided
     - Check the availability of environment variables
     - Hostcode validation is available from previous commits

  - Creation of a separate module/utils file for passwordless SSH
    setup. This will aid standalone execution of a passwordless SSH
    setup process.

  - Addition of a passwordless SSH optional switch --passwordless
    which takes values True/False. Running the script with this
    switch set to True will cause skip addition of ansible_ssh_password
    hostvar in the inventory hosts file. Default value is False.

  - Addition of comment blocks in generate-inventory.py and passwordless.py
    to provide purpose and run-instructions.

- Modified the run-iometer-demo playbook to remove section where Dockerfile
  path was being sent as an extra var/args.

- Updated the openebs-on-premise-deployment-guide.md's prerequisites
  section to add instructions about env variable setup

- Updated README.md for ansible roles to reflect changes in inventory
  generation
