+++ e2e/ansible/files/README.md
@@ -1,5 +1,4 @@
 # Ansible Files
 ---------------
 Contains auxiliary files such as images, scripts, etc., which may be used during a playbook run. 
-Currently Includes python scripts to generate inventory file.
