+++ e2e/ansible/roles/inventory/defaults/main.yml
@@ -1,4 +1,2 @@
 ---
-inventory_script_path: "{{hostvars['localhost'].playbook_dir}}/files"
-
 inventory_input_path: "{{hostvars['localhost'].playbook_dir}}/inventory"
