+++ e2e/ansible/roles/iometer/files/Dockerfile
